Removed reference to quadtree on the quadtree nodes. Deleted old unused code.

+++ app/src/main/java/com/sdgapps/terrainsandbox/MiniEngine/terrain/Planet.java
@@ -81,7 +81,7 @@ public class Planet extends Renderer implements TerrainInterface {
 
     private int gridSize = 64;
     private float rootQuadScale = 100000;
-    private int nLods = 8;
+    private short nLods = 8;
     private float yscale = 60000;
     public float terrainXZ;
